getting tokyonight theme working
I had trouble initially - the package was not found even though it
suggested that it was installed. I had the same issue with the whichkey
plugin before - after adding it as a plugin to download in my plugins
folder lazy.nvim will automatically download the plugin, but I need to
create a file somewhere so I can call
`require("tokyonight").setup({...})` and then I can add the file that
has this call the my `init.lua` file. If no file requires/setsup/configs
and gets the preferences for the plugin we want to use, then lazy.nvim
will not actually load that plugin, so it will not be available. This
might be overkill, but for *every plugin* we should have a *config file*
that sets it up and sets keybindings and other preferences