TELESCOPE PLUGIN IS AMAZING
I added telescope, which does a couple things - it fuzzy finds files
recursively *from the directory that we are currently in*. We can also
search for text patterns (equivalent of `grep`). We can open recent
files. This makes navigating between files when you know what you want a
lot easier. I still need to have like goto definition and goto
declaration and go from one file to another seamlessly, but this is a
step closer. You can also search for commands that I have existing in
neovim using the search, and I can change colorschemes in two seconds
which is easy when going from something inside to solarized outside