adding indentlines in plugins
this adds the vertical lines to make the tabs easier to see. There is
also an option to add something to show where enters (carriage returns)
are in the code too. I think that I need to go back to nvimtree and
figure out why the icons are not showing. That and some of the options
that I set to indent text do not seem to be working, but I think
everything else is