adding lazygit to neovim
added lazy git and figured out where in whichkey to put things. The indentation of a character means that when we run which key we get another list of characters and this can be done recursively. Also, I can search for which commands to run using Treesitter and then I can run <cmd> which represents pressing a `:` and then type the command and add a carriage return <cr> to represent pressing enter. This is the first plugin that I added completely by myself